# docs/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/notifications/me/unread-count:
    get:
      operationId: NotificationsController_unreadCount
      parameters: []
      responses:
        '200':
          description: Unread count.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponseDto'
      security: &a47
        - bearer: []
      summary: 'Returns { unread: N } for the home-screen badge. Single indexed COUNT
        — cheap.'
      tags: &a48
        - notifications
  /api/v1/notifications/me:
    get:
      operationId: NotificationsController_listMine
      parameters:
        - name: limit
          required: true
          in: query
          schema:
            type: string
        - name: channel
          required: true
          in: query
          schema:
            type: string
        - name: includeArchived
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Caller's recent deliveries, newest-first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMyNotificationsResponseDto'
        '400':
          description: VALIDATION_FAILED — channel is not one of push|email|sms.
      security: *a47
      summary: List the caller's recent notifications. Optional
        ?channel=push|email|sms narrows to one delivery channel.
      tags: *a48
  /api/v1/notifications/read-all:
    post:
      operationId: NotificationsController_markAllRead
      parameters: []
      responses:
        '200':
          description: Number of rows actually flipped.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllReadResponseDto'
      security: *a47
      summary: 'Mark every unread row as read. Returns { marked: N }. Idempotent
        (second call returns 0).'
      tags: *a48
  /api/v1/notifications/{id}/read:
    post:
      operationId: NotificationsController_markRead
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Updated notification row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationLogDto'
        '404':
          description: NOTIFICATION_NOT_FOUND.
      security: *a47
      summary: Mark one notification as read. Owner-gated; 404 on cross-user / missing.
      tags: *a48
  /api/v1/notifications/{id}/unread:
    post:
      operationId: NotificationsController_markUnread
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Updated notification row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationLogDto'
        '404':
          description: NOTIFICATION_NOT_FOUND.
      security: *a47
      summary: Mark one notification as UNread (symmetric to /read). Owner-gated.
      tags: *a48
  /api/v1/notifications/{id}:
    delete:
      operationId: NotificationsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '404':
          description: NOTIFICATION_NOT_FOUND.
      security: *a47
      summary: Hard-delete a notification (inbox prune). Owner-gated; 404 on
        cross-user / missing.
      tags: *a48
  /api/v1/notifications/{id}/archive:
    post:
      operationId: NotificationsController_archive
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Archived.
        '404':
          description: NOTIFICATION_NOT_FOUND.
      security: *a47
      summary: Soft-archive a notification (swipe-to-archive). Owner-gated; 404 on
        cross-user / missing. Idempotent.
      tags: *a48
...
  /api/v1/notifications/push/subscribe:
    post:
      operationId: PushSubscriptionsController_subscribe
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribePushRequestDto'
      responses:
        '201':
          description: Subscription stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushSubscriptionDto'
        '422':
          description: VALIDATION_FAILED.
      security: &a51
        - bearer: []
      summary: Register or refresh a Web Push subscription. Idempotent on endpoint —
        re-subscribing the same browser takes ownership.
      tags: &a52
        - notifications
    delete:
      operationId: PushSubscriptionsController_unsubscribe
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscribePushRequestDto'
      responses:
        '204':
          description: Unsubscribed.
        '404':
          description: PUSH_SUBSCRIPTION_NOT_FOUND.
      security: *a51
      summary: Owner-scoped Web Push unsubscribe. 404 if the endpoint is not
        registered to caller.
      tags: *a52
...
  /api/v1/reviews/summary:
    get:
      operationId: ReviewsController_summary
      parameters:
        - name: targetType
          required: true
          in: query
          schema:
            type: string
        - name: targetId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Summary; zero-filled for empty targets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSummaryDto'
        '400':
          description: VALIDATION_FAILED — targetType / targetId missing or invalid.
      security: *a59
      summary: Aggregated review summary { count, average, histogram } for a target.
        @Public.
      tags: *a60
...
info:
  title: TravelSuperApp API
  description: OpenAPI export from the running NestJS app. Per-route schemas land
    progressively as @ApiBody / @ApiResponse decorators get added module by
    module — see docs/api/README.md.
  version: v1
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: 'Access token issued by POST /auth/login or rotated via POST
        /auth/refresh. Lifetime: 15 minutes (configurable via
        JWT_ACCESS_EXPIRY).'
  schemas:
...
    UnreadCountResponseDto:
      type: object
      properties:
        unread:
          type: number
          description: Count of unread rows across all channels.
      required:
        - unread
    NotificationLogDto:
      type: object
      properties:
        id:
          type: string
          format: cuid
        channel:
          type: string
          enum:
            - push
            - email
            - sms
        templateId:
          type: string
          description: Template id (e.g. "trip.invited", "sos.acknowledged").
        status:
          type: string
          enum:
            - queued
            - delivered
            - failed
            - skipped
          description: Lifecycle status of the delivery attempt.
        payload:
          type: object
          additionalProperties: true
          description: Template payload (JSON). Shape varies per template id.
        read:
          type: boolean
          description: True once the recipient has read the notification.
        archivedAt:
          type: object
          nullable: true
          format: date-time
          description: V.UX.26 — set when the user swipe-archived the row. Default lister
            filters non-null out.
        createdAt:
          type: string
          format: date-time
        deliveredAt:
          type: object
          nullable: true
          format: date-time
          description: When the deliver-side adapter reported success; null while
            queued/failed.
      required:
        - id
        - channel
        - templateId
        - status
        - payload
        - read
        - archivedAt
        - createdAt
        - deliveredAt
    ListMyNotificationsResponseDto:
      type: object
      properties:
        notifications:
          description: Caller's recent deliveries, newest-first. Capped via ?limit
            (1..200, default 50).
          type: array
          items:
            $ref: '#/components/schemas/NotificationLogDto'
      required:
        - notifications
    MarkAllReadResponseDto:
      type: object
      properties:
        marked:
          type: number
          description: Count of rows actually flipped on this call. 0 on idempotent re-call.
      required:
        - marked
...
    SubscribePushKeysDto:
      type: object
      properties:
        p256dh:
          type: string
          description: Web Push spec §3 — base64 client public key.
        auth:
          type: string
          description: Web Push spec §3 — base64 auth secret.
      required:
        - p256dh
        - auth
    SubscribePushRequestDto:
      type: object
      properties:
        endpoint:
          type: string
          description: Browser-vendor endpoint URL.
        keys:
          $ref: '#/components/schemas/SubscribePushKeysDto'
      required:
        - endpoint
        - keys
    PushSubscriptionDto:
      type: object
      properties:
        id:
          type: string
          format: cuid
        userId:
          type: string
          format: cuid
        endpoint:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - endpoint
        - createdAt
    UnsubscribePushRequestDto:
      type: object
      properties:
        endpoint:
          type: string
          description: Browser-vendor endpoint URL to unsubscribe.
      required:
        - endpoint
...
    ReviewSummaryDto:
      type: object
      properties:
        targetType:
          type: string
          enum:
            - place
            - stay
            - eatery
            - agent
        targetId:
          type: string
        count:
          type: number
          description: Number of reviews on the target.
        average:
          type: number
          description: Mean rating, two-decimal-rounded.
        histogram:
          type: object
          additionalProperties:
            type: number
          description: Histogram keyed by rating bucket ("1".."5") → count.
      required:
        - targetType
        - targetId
        - count
        - average
        - histogram
...